Refactor reducers into logical units (kind of)
