+++ web/src/reducers/index.js
@@ -1,185 +1,17 @@
 import updeep from 'updeep';
 import shape from './shape/';
-import { IAPD, Projects, Requests } from '../actions';
+import { Projects, Requests } from '../actions';
 
-const randomID = (function createRandomIDFunction() {
-  const letters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'.split('');
-  return () => {
-    let id = '';
-    for (let i = 0; i < 8; i += 1) {
-      id += letters[Math.floor(Math.random() * letters.length)];
-    }
-    return id;
-  };
-}());
+import iapdReducer from './iapd';
 
 const stateShape = shape.app;
 
 export default function reducer(state = stateShape, action) {
   let newState = state;
 
+  newState = updeep({ currentRequest: iapdReducer(state.currentRequest, action) }, newState);
+
   switch (action.type) {
-    case IAPD.messages.UPDATE_EXECUTIVE_SUMMARY:
-      newState = updeep({ currentRequest: { executiveSummary: action.executiveSummary } }, newState);
-      break;
-
-    case IAPD.messages.UPDATE_PROJECT_PLAN:
-      newState = updeep({ currentRequest: { plan: action.projectPlan } }, newState);
-      break;
-
-    case IAPD.messages.UPDATE_PROJECT_MILESTONES:
-      newState = updeep({ currentRequest: { milestones: action.milestones } }, newState);
-      break;
-
-    case IAPD.messages.UPDATE_BUDGET_DDI_CATEGORY:
-      {
-        const categoryUpdate = { [action.category]: action.values };
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { ddi: { categories: categoryUpdate } } } } }, newState);
-
-        // Update DDI totals too
-        let ffp90 = 0;
-        let state10 = 0;
-        let ffp75 = 0;
-        let state25 = 0;
-        let total = 0;
-
-        for (const category of newState.currentRequest.budget[action.budget].ddi.categories) {
-          ffp90 += category.ffp90;
-          state10 += category.state10;
-          ffp75 += category.ffp75;
-          state25 += category.state25;
-          total += category.total;
-        }
-
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { ddi: { ffp90, state10, ffp75, state25, total } } } } }, newState);
-      }
-      break;
-
-    case IAPD.messages.UPDATE_BUDGET_OM_CATEGORY:
-      {
-        const categoryUpdate = { [action.category]: action.values };
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { om: { categories: categoryUpdate } } } } }, newState);
-
-        // Update O&M totals too
-        let ffp75 = 0;
-        let state25 = 0;
-        let ffp50 = 0;
-        let state50 = 0;
-        let total = 0;
-
-        for (const category of newState.currentRequest.budget[action.budget].om.categories) {
-          ffp75 += category.ffp75;
-          state25 += category.state25;
-          ffp50 += category.ffp50;
-          state50 += category.state50;
-          total += category.total;
-        }
-
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { om: { ffp75, state25, ffp50, state50, total } } } } }, newState);
-      }
-      break;
-
-    case IAPD.messages.UPDATE_BUDGET_MECH_CATEGORY:
-      {
-        const categoryUpdate = { [action.category]: action.values };
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { mechanizedSystems: { categories: categoryUpdate } } } } }, newState);
-
-        // Update mechanized systems totals too
-        let ffp50 = 0;
-        let state50 = 0;
-        let total = 0;
-
-        for (const category of newState.currentRequest.budget[action.budget].mechanizedSystems.categories) {
-          ffp50 += category.ffp50;
-          state50 += category.state50;
-          total += category.total;
-        }
-
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { mechanizedSystems: { ffp50, state50, total } } } } }, newState);
-      }
-      break;
-
-    case IAPD.messages.UPDATE_PAPD_SUMMARY:
-      newState = updeep({ currentRequest: { papdSummary: action.papdSummary } }, newState);
-      break;
-
-    case IAPD.messages.ADD_PROJECT_OUTCOME:
-      {
-        const outcomes = [...newState.currentRequest.plan.outcomes];
-        outcomes.push({
-          id: randomID(),
-          title: '',
-          metrics: ''
-        });
-        newState = updeep({ currentRequest: { plan: { outcomes } } }, newState);
-      }
-      break;
-
-    case IAPD.messages.ADD_PROJECT_MILESTONE:
-      {
-        const milestones = [...newState.currentRequest.milestones];
-        milestones.push({
-          id: randomID(),
-          description: '',
-          activities: '',
-          associatedOutcomes: [],
-          mitaAreas: [],
-          cost: 0,
-          defineSuccess: ''
-        });
-        newState = updeep({ currentRequest: { milestones } }, newState);
-      }
-      break;
-
-    case IAPD.messages.ADD_BUDGET_DDI_CATEGORY:
-      {
-        const categories = [...newState.currentRequest.budget[action.budget].ddi.categories];
-        categories.push({
-          id: randomID(),
-          category: '',
-          inhouse: null,
-          ffp90: 0,
-          state10: 0,
-          ffp75: 0,
-          state25: 0,
-          total: 0
-        });
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { ddi: { categories } } } } }, newState);
-      }
-      break;
-
-    case IAPD.messages.ADD_BUDGET_OM_CATEGORY:
-      {
-        const categories = [...newState.currentRequest.budget[action.budget].om.categories];
-        categories.push({
-          id: randomID(),
-          category: '',
-          inhouse: null,
-          ffp75: 0,
-          state25: 0,
-          ffp50: 0,
-          state50: 0,
-          total: 0
-        });
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { om: { categories } } } } }, newState);
-      }
-      break;
-
-    case IAPD.messages.ADD_BUDGET_MECH_CATEGORY:
-      {
-        const categories = [...newState.currentRequest.budget[action.budget].mechanizedSystems.categories];
-        categories.push({
-          id: randomID(),
-          category: '',
-          inhouse: null,
-          ffp50: 0,
-          state50: 0,
-          total: 0
-        });
-        newState = updeep({ currentRequest: { budget: { [action.budget]: { mechanizedSystems: { categories } } } } }, newState);
-      }
-      break;
-
     case Requests.messages.START_NEW_REQUEST:
       newState = updeep({ currentRequest: shape.request }, newState);
       break;
